v1: `PreciseFlex`: name the error module `errors.py` and co-locate `PreciseFlexError`
`errors.py` is the codebase convention for a backend's error module - the Hamilton STAR
`errors.py` holds both the error-code tables and the exception classes. This renames the
PreciseFlex code-table module from `error_codes.py` to `errors.py` and moves `PreciseFlexError`
(which reads `ERROR_CODES`) into it, so the table and the exception that consumes it live
together.

No behaviour change. The public import paths are unchanged: `PreciseFlexError` is still
re-exported from the package `__init__` and remains importable from `precise_flex`.
